requirements.txt
streamlit 
requests 
pymongo
python-dotenv